build: Fix building clboss as a git submodule
When clboss is a git submodule, the current Makefile fails because
it assumes that because the `$(srcdir)/.git/logs/HEAD` file does
not exist, that we must be inside the release zip and therefore
already have the commit_hash.h file present.

Here we update the Makefile to detect the case where clboss is
a git submodule (by detecting that `.git` is a regular file and
not a directory), and run generate_commit_hash.sh in that case.